@@ -2,7 +2,7 @@
 name         = "esp-metadata"
 version      = "0.3.0"
 edition      = "2021"
-rust-version = "1.60.0"
+rust-version = "1.70.0"
 description  = "Metadata for Espressif devices"
 repository   = "https://github.com/esp-rs/esp-hal"
 license      = "MIT OR Apache-2.0"
@@ -11,6 +11,5 @@ license      = "MIT OR Apache-2.0"
 anyhow     = "1.0.86"
 clap       = { version = "4.5.16", features = ["derive"] }
 basic-toml  = "0.1.9"
-lazy_static = "1.5.0"
 serde       = { version = "1.0.209", features = ["derive"] }
 strum       = { version = "0.26.3",  features = ["derive"] }
